python-service/requirements.txt
# Python Analytics Service Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1

# Database (for real data integration)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Science & Analytics (for bi_analytics_ai.py)
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
